Start working on training script and finalizing new emulator
