+++ summit/benchmarks/experimental_emulator.py
@@ -1,160 +1,158 @@
-import os
-import os.path as osp
-
-from summit.experiment import Experiment
-
-import numpy as np
-
-from summit.benchmarks.experiment_emulator.bnn_emulator import BNNEmulator
 from summit.utils.dataset import DataSet
 from summit.domain import *
-from summit.utils import jsonify_dict, unjsonify_dict
+from summit.experiment import Experiment
+from summit import get_summit_config_path
+from summit.strategies import Transform
+
+import torch
+import pytorch_lightning as pl
+
+from blitz.modules import BayesianLinear
+from blitz.utils import variational_estimator
+
+from sklearn.preprocessing import StandardScaler
+from sklearn.model_selection import train_test_split
+
+import matplotlib.pyplot as plt
+import pathlib
+import numpy as np
+from numpy.random import default_rng
 
 
 class ExperimentalEmulator(Experiment):
     """Experimental Emulator
 
+    Train a machine learning model based on experimental data.
+    The model acts a benchmark for testing optimisation strategies.
+
     Parameters
-    ---------
-    domain: summit.domain.Domain
-        The domain of the experiment
-    dataset: class:~summit.utils.dataset.DataSet, optional
-        A DataSet with data for training where the data columns correspond to the domain and the data rows correspond to the training points.
-        By default: None
-    csv_dataset: string, optional
-        Path to csv_file with data for training where columns correspond to the domain and the rows correspond to the training points.
-        Note that the first row should exactly match the variable names of the domain and the second row should only have "DATA" as entry.
-        By default: None
-    model_name: string, optional
-        Name of the model that is used for saving model parameters. Should be unique.
-        By default: "dataset_emulator_model_name"
-    regressor_type: string, optional
-        Type of the regressor that is used within the emulator (available: "BNN").
-        By default: "BNN"
-    cat_to_descr: Boolean, optional
-        If True, transform categorical variable to one or more continuous variable(s)
-        corresponding to the descriptors of the categorical variable (else do nothing).
-        By default: False
+    ----------
+    model_name : str
+        Model name used for identification. Must be unique from other models stored on the system.
+    domain : :class:`~summit.domain.Domain`
+        The domain of the emulator
+    dataset : :class:`~summit.dataset.Dataset`
+        Dataset used for training/validation
+    regressor : :classs:`pl.LightningModule`, optional
+        Pytorch LightningModule class. Defaults to the BayesianRegressor
+    model_dir : :class:`pathlib.Path` or str, optional
+        Directory where models are saved. Defaults to `~/.summit/ExperimentEmulator"
+    load_checkpoint : bool, optional
+        Whether to load any previously trained models on disk. By default previous models are not loaded.
+    normalize : bool, optional
+        Normalize continuous input variables. Default is True.
+    test_size : float, optional
+        Fraction of data used for test. Default is 0.25
+    random_state : float, optional
+        A random initialization value. Use to make results more reproducible.
+
+
+    Raises
+    ------
+    ValueError
+        description
 
     Examples
     --------
-    >>> test_domain = ReizmanSuzukiEmulator().domain
-    >>> e = ExperimentalEmulator(domain=test_domain, model_name="Pytest")
-    No trained model for Pytest. Train this model with ExperimentalEmulator.train() in order to use this Emulator as an virtual Experiment.
-    >>> columns = [v.name for v in e.domain.variables]
-    >>> train_values = {("catalyst", "DATA"): ["P1-L2", "P1-L7", "P1-L3", "P1-L3"], ("t_res", "DATA"): [60, 120, 110, 250], ("temperature", "DATA"): [110, 30, 70, 80], ("catalyst_loading", "DATA"): [0.508, 0.6, 1.4, 1.3], ("yield", "DATA"): [20, 40, 60, 34], ("ton", "DATA"): [33, 34, 21, 22]}
-    >>> train_dataset = DataSet(train_values, columns=columns)
-    >>> e.train(train_dataset, verbose=False, cv_fold=2, test_size=0.25)
-    >>> columns = [v.name for v in e.domain.variables]
-    >>> values = [float(v.bounds[0] + 0.6 * (v.bounds[1] - v.bounds[0])) if v.variable_type == 'continuous' else v.levels[-1] for v in e.domain.variables]
-    >>> values = np.array(values)
-    >>> values = np.atleast_2d(values)
-    >>> conditions = DataSet(values, columns=columns)
-    >>> results = e.run_experiments(conditions)
+    >>>
+    Notes
+    -----
 
     """
 
-    def __init__(
-        self,
-        domain,
-        dataset=None,
-        csv_dataset=None,
-        model_name="dataset_name_emulator_bnn",
-        regressor_type="BNN",
-        cat_to_descr=False,
-        **kwargs
-    ):
-        super().__init__(domain)
-        dataset = self._check_datasets(dataset, csv_dataset)
+    def __init__(self, model_name, domain, dataset=None, **kwargs):
+        super().__init__(domain, **kwargs)
 
-        kwargs["cat_to_descr"] = cat_to_descr
+        # Save locations
+        self.model_name = model_name
+        self.model_dir = kwargs.get(
+            "model_dir", get_summit_config_path() / "ExprimentalEmulator"
+        )
+        self.model_dir = pathlib.Path(self.model_dir)
+        self.checkpoint_path = (
+            self.model_dir / self.model_name / f"{self.model_name}.ckpt"
+        )
 
-        if regressor_type == "BNN":
-            self.emulator = BNNEmulator(
-                domain=domain, dataset=dataset, model_name=model_name, kwargs=kwargs
+        # Create the datamoedule
+        if dataset is not None:
+            self.datamodule = EmulatorDataModule(self.domain, dataset, **kwargs)
+
+        # Create the regressor
+        self.regressor = kwargs.get("regressor")
+        if not self.regressor and dataset is not None:
+            train_loader = self.datamodule.train_dataloader()
+            self.n_examples = len(train_loader.dataset)
+            n_features = train_loader.dataset[0][0].shape[0]
+            n_targets = train_loader.dataset[0][1].shape[0]
+            self.regressor = BayesianRegressor(n_features, n_targets, self.n_examples)
+
+        # Try to load any previous models
+        load_checkpoint = kwargs.get("load_checkpoint", False)
+        if self.checkpoint_path.exists() and load_checkpoint:
+            self.regressor = BayesianRegressor.load_from_checkpoint(
+                checkpoint_path=self.checkpoint_path
             )
-            try:
-                self.extras = [self.emulator._load_model(model_name)]
-            except:
-                print(
-                    "No trained model for {}. Train this model with ExperimentalEmulator.train() in order to use this Emulator as an virtual Experiment.".format(
-                        self.emulator.model_name
-                    )
-                )
-        else:
-            raise NotImplementedError(
-                "Regressor type <{}> not implemented yet".format(str(regressor_type))
+            print("Model Loaded from disk")
+        elif dataset is None:
+            raise ValueError(
+                "Dataset cannot be None when there is not pretrained model."
             )
 
     def _run(self, conditions, **kwargs):
-        condition = DataSet.from_df(conditions.to_frame().T)
-        infer_dict = self.emulator.infer_model(dataset=condition)
-        for k, v in infer_dict.items():
-            conditions[(k, "DATA")] = v
-        return conditions, None
+        X = conditions[[v.name for v in self.domain.input_variables]]
+        if self.datamodule.normalize:
+            X = self.datamodule.input_scaler.transform(X)
+        y = self.regressor(X)
+        if self.datamodule.normalize:
+            y = self.datamodule.output_scaler.inverse_transform(y)
+        for i, v in enumerate(self.domain.output_variables):
+            conditions[v.name] = y[:, i]
+        return conditions
 
-    def train(self, dataset=None, csv_dataset=None, verbose=True, **kwargs):
-        dataset = self._check_datasets(dataset, csv_dataset)
-        self.emulator.set_training_hyperparameters(kwargs=kwargs)
-        self.emulator.train_model(dataset=dataset, verbose=verbose, kwargs=kwargs)
-        self.extras = [self.emulator.output_models]
+    def train(self, **kwargs):
+        """Train the model on the dataset
 
-    def validate(
-        self,
-        dataset=None,
-        csv_dataset=None,
-        parity_plots=False,
-        get_pred=False,
-        **kwargs
-    ):
-        dataset = self._check_datasets(dataset, csv_dataset)
-        if dataset is not None:
-            return self.emulator.validate_model(
-                dataset=dataset,
-                parity_plots=parity_plots,
-                get_pred=get_pred,
-                kwargs=kwargs,
-            )
-        else:
-            try:
-                print("Evaluation based on training and test set.")
-                return self.emulator.validate_model(parity_plots=parity_plots)
-            except:
-                raise ValueError("No dataset to evaluate.")
-
-    def _check_datasets(self, dataset=None, csv_dataset=None):
-        if csv_dataset:
-            if dataset:
-                print(
-                    "Dataset and csv.dataset are given, hence dataset will be overwritten by csv.data."
-                )
-            dataset = DataSet.read_csv(csv_dataset, index_col=None)
-        return dataset
-
-    def to_dict(self, **kwargs):
-        """Serialize the class to a dictionary
-
-        Subclasses can add a experiment_params dictionary
-        key with custom parameters for the experiment
+        For kwargs, see pytorch-lightining documentation:
+        https://pytorch-lightning.readthedocs.io/en/stable/trainer.html#init
         """
-        kwargs.update(
-            dict(
-                model_name=self.emulator.model_name,
-                dataset=self.emulator._dataset.to_dict()
-                if self.emulator._dataset is not None
-                else None,
-                output_models=self.emulator.output_models,
+        logger = kwargs.get("logger")
+        version = kwargs.get("version", 0)
+        if logger is None:
+            kwargs["logger"] = pl.loggers.TensorBoardLogger(
+                name=self.model_name,
+                save_dir=self.model_dir,
+                version=version,
             )
-        )
-        return super().to_dict(**kwargs)
+        kwargs["checkpoint_callback"] = kwargs.get("checkpoint_callback", False)
 
-    @classmethod
-    def from_dict(cls, d):
-        dataset = d["experiment_params"]["dataset"]
-        d["experiment_params"]["dataset"] = DataSet.from_dict(dataset)
-        exp = super().from_dict(d)
-        exp.emulator.output_models = d["experiment_params"]["output_models"]
-        return exp
+        # Use pytorch lightining for training and saving
+        trainer = pl.Trainer(**kwargs)
+        trainer.fit(self.regressor, self.datamodule)
+        trainer.save_checkpoint(self.checkpoint_path)
+
+    def to_dict(self):
+        params = dict(
+            dataset=self.dataset,
+            model_name=self.model_name,
+            model_dir=self.model_dir,
+            n_examples=self.n_examples,
+        )
+        return super().to_dict(**params)
+
+    def from_dict(self):
+        pass
+
+    def parity_plot(self):
+        """ Produce a parity plot based on the test data"""
+        X_test, y_test = self.datamodule.X_test, self.datamodule.y_test
+        with torch.no_grad():
+            Y_test_pred = self.regressor(X_test)
+        fig, ax = plt.subplots(1)
+        ax.scatter(y_test[:, 0], Y_test_pred[:, 0])
+        ax.plot([-4, 4], [-4, 4])
+        ax.set_xlabel("Measured")
+        ax.set_ylabel("Predicted")
+        return fig, ax
 
 
 class ReizmanSuzukiEmulator(ExperimentalEmulator):
@@ -186,14 +184,10 @@ class ReizmanSuzukiEmulator(ExperimentalEmulator):
 
     """
 
-    def __init__(self, case=1, **kwargs):
-        model_name = "reizman_suzuki_case" + str(case)
+    def __init__(self, case=1, dataset=None, **kwargs):
+        model_name = f"reizman_suzuki_case_{case}"
         domain = self.setup_domain()
-        dataset_file = osp.join(
-            osp.dirname(osp.realpath(__file__)),
-            "experiment_emulator/data/" + model_name + "_train_test.csv",
-        )
-        super().__init__(domain=domain, model_name=model_name)
+        super().__init__(model_name, domain, dataset=dataset, **kwargs)
 
     def setup_domain(self):
         domain = Domain()
@@ -251,335 +245,285 @@ class ReizmanSuzukiEmulator(ExperimentalEmulator):
 
         return domain
 
-    def to_dict(self):
-        """Serialize the class to a dictionary"""
-        experiment_params = dict(
-            case=self.emulator.model_name[-1],
-        )
-        return super().to_dict(**experiment_params)
 
-
-class BaumgartnerCrossCouplingEmulator(ExperimentalEmulator):
-    """Baumgartner Cross Coupling Emulator
-
-    Virtual experiments representing the Aniline Cross-Coupling reaction
-    similar to Baumgartner et al. (2019). Experimental outcomes are based on an
-    emulator that is trained on the experimental data published by Baumgartner et al.
-
-    This is a five dimensional optimisation of temperature, residence time, base equivalents,
-    catalyst and base.
-
-    The categorical variables (catalyst and base) contain descriptors
-    calculated using COSMO-RS. Specifically, the descriptors are the first two sigma moments.
-
-    Examples
-    --------
-    >>> bemul = BaumgartnerCrossCouplingDescriptorEmulator()
-
-    Notes
-    -----
-    This benchmark is based on data from [Baumgartner]_ et al.
-
-    References
-    ----------
-
-    .. [Baumgartner] L. M. Baumgartner et al., Org. Process Res. Dev., 2019, 23, 1594–1601
-       DOI: `10.1021/acs.oprd.9b00236 <https://`doi.org/10.1021/acs.oprd.9b00236>`_
-
-    """
-
-    def __init__(self, **kwargs):
-        model_name = kwargs.get("model_name", "baumgartner_aniline_cn_crosscoupling")
-        dataset_file = kwargs.get(
-            "dataset_file", "baumgartner_aniline_cn_crosscoupling.csv"
-        )
-        domain = self.setup_domain()
-        dataset_file = osp.join(
-            osp.dirname(osp.realpath(__file__)),
-            "experiment_emulator/data/" + dataset_file,
-        )
-        super().__init__(domain=domain, csv_dataset=dataset_file, model_name=model_name)
-
-    def setup_domain(self):
-        domain = Domain()
-
-        # Decision variables
-        des_1 = "Catalyst type"
-        catalyst_df = DataSet(
-            [
-                [460.7543, 67.2057],  # 30.8413, 2.3043, 0], #, 424.64, 421.25040226],
-                [518.8408, 89.8738],  # 39.4424, 2.5548, 0], #, 487.7, 781.11247064],
-                [819.933, 129.0808],  # 83.2017, 4.2959, 0], #, 815.06, 880.74916884],
-            ],
-            index=["tBuXPhos", "tBuBrettPhos", "AlPhos"],
-            columns=[
-                "area_cat",
-                "M2_cat",
-            ],  # , 'M3_cat', 'Macc3_cat', 'Mdon3_cat'] #,'mol_weight', 'sol']
-        )
-        domain += CategoricalVariable(
-            name="catalyst",
-            description=des_1,
-            levels=["tBuXPhos", "tBuBrettPhos", "AlPhos"],
-            descriptors=catalyst_df,
-        )
-
-        des_2 = "Base"
-        base_df = DataSet(
-            [
-                [162.2992, 25.8165],  # 40.9469, 3.0278, 0], #101.19, 642.2973283],
-                [
-                    165.5447,
-                    81.4847,
-                ],  # 107.0287, 10.215, 0.0169], # 115.18, 534.01544123],
-                [227.3523, 30.554],  # 14.3676, 1.1196, 0.0127], # 171.28, 839.81215],
-                [192.4693, 59.8367],  # 82.0661, 7.42, 0], # 152.24, 1055.82799],
-            ],
-            index=["TEA", "TMG", "BTMG", "DBU"],
-            columns=["area", "M2"],  # , 'M3', 'Macc3', 'Mdon3'], # 'mol_weight', 'sol']
-        )
-        domain += CategoricalVariable(
-            name="base",
-            description=des_2,
-            levels=["DBU", "BTMG", "TMG", "TEA"],
-            descriptors=base_df,
-        )
-
-        des_3 = "Base equivalents"
-        domain += ContinuousVariable(
-            name="base_equivalents", description=des_3, bounds=[1.0, 2.5]
-        )
-
-        des_4 = "Temperature in degrees Celsius (ºC)"
-        domain += ContinuousVariable(
-            name="temperature", description=des_4, bounds=[30, 100]
-        )
-
-        des_5 = "residence time in seconds (s)"
-        domain += ContinuousVariable(name="t_res", description=des_5, bounds=[60, 1800])
-
-        des_6 = "Yield"
-        domain += ContinuousVariable(
-            name="yld",
-            description=des_6,
-            bounds=[0.0, 1.0],
-            is_objective=True,
-            maximize=True,
-        )
-
-        return domain
-
-
-class BaumgartnerCrossCouplingDescriptorEmulator(ExperimentalEmulator):
-    """Baumgartner Cross Coupling Emulator
-
-    Virtual experiments representing the Aniline Cross-Coupling reaction
-    similar to Baumgartner et al. (2019). Experimental outcomes are based on an
-    emulator that is trained on the experimental data published by Baumgartner et al.
-
-    The difference with this model is that it uses descriptors for the catalyst and base
-    instead of one-hot encoding the options. The descriptors are the first two
-    sigma moments from COSMO-RS.
-
+class EmulatorDataModule(pl.LightningDataModule):
+    """Convert Summit DataSet and Domain into a Pytorch Dataloader
 
     Parameters
     ----------
+    domain : :class:`~summit.domain.Domain`
+        Domain of the optimization
+    dataset : :class:`~summit.dataset.Dataset`
+        Dataset used for training/validation
+    normalize : bool, optional
+        Normalize continuous input and output variables. Default is True.
+    test_size : float, optional
+        Fraction of data used for test. Default is 0.25
+    random_state : float, optional
+        A random initialization value. Use to make results more reproducible.
+
+    Returns
+    -------
+    result: `bool`
+        description
+
+    Raises
+    ------
+    ValueError
+        description
 
     Examples
     --------
-    >>> bemul = BaumgartnerCrossCouplingDescriptorEmulator()
+
 
     Notes
     -----
-    This benchmark is based on data from [Baumgartner]_ et al.
 
-    References
-    ----------
-
-    .. [Baumgartner] L. M. Baumgartner et al., Org. Process Res. Dev., 2019, 23, 1594–1601
-       DOI: `10.1021/acs.oprd.9b00236 <https://doi.org/10.1021/acs.oprd.9b00236>`_
 
     """
 
-    def __init__(self, **kwargs):
-        model_name = kwargs.get(
-            "model_name", "baumgartner_aniline_cn_crosscoupling_descriptors"
-        )
-        dataset_file = kwargs.get(
-            "dataset_file", "baumgartner_aniline_cn_crosscoupling_descriptors.csv"
-        )
-        domain = self.setup_domain()
-        dataset_file = osp.join(
-            osp.dirname(osp.realpath(__file__)),
-            "experiment_emulator/data/" + dataset_file,
-        )
-        super().__init__(domain=domain, csv_dataset=dataset_file, model_name=model_name)
-
-    def setup_domain(self):
-        domain = Domain()
-
-        # Decision variables
-        des_1 = "Catalyst type with descriptors"
-        catalyst_df = DataSet(
-            [
-                [460.7543, 67.2057, 30.8413, 2.3043, 0],  # , 424.64, 421.25040226],
-                [518.8408, 89.8738, 39.4424, 2.5548, 0],  # , 487.7, 781.11247064],
-                [819.933, 129.0808, 83.2017, 4.2959, 0],  # , 815.06, 880.74916884],
-            ],
-            index=["tBuXPhos", "tBuBrettPhos", "AlPhos"],
-            columns=[
-                "area_cat",
-                "M2_cat",
-                "M3_cat",
-                "Macc3_cat",
-                "Mdon3_cat",
-            ],  # ,'mol_weight', 'sol']
-        )
-        domain += CategoricalVariable(
-            name="catalyst", description=des_1, descriptors=catalyst_df
-        )
-
-        des_2 = "Base type with descriptors"
-        base_df = DataSet(
-            [
-                [162.2992, 25.8165, 40.9469, 3.0278, 0],  # 101.19, 642.2973283],
-                [165.5447, 81.4847, 107.0287, 10.215, 0.0169],  # 115.18, 534.01544123],
-                [227.3523, 30.554, 14.3676, 1.1196, 0.0127],  # 171.28, 839.81215],
-                [192.4693, 59.8367, 82.0661, 7.42, 0],  # 152.24, 1055.82799],
-            ],
-            index=["TEA", "TMG", "BTMG", "DBU"],
-            columns=["area", "M2", "M3", "Macc3", "Mdon3"],  # 'mol_weight', 'sol']
-        )
-        domain += CategoricalVariable(
-            name="base", description=des_2, descriptors=base_df
-        )
-
-        des_3 = "Base equivalents"
-        domain += ContinuousVariable(
-            name="base_equivalents", description=des_3, bounds=[1.0, 2.5]
-        )
-
-        des_4 = "Temperature in degrees Celsius (ºC)"
-        domain += ContinuousVariable(
-            name="temperature", description=des_4, bounds=[30, 100]
-        )
-
-        des_5 = "residence time in seconds (s)"
-        domain += ContinuousVariable(name="t_res", description=des_5, bounds=[60, 1800])
-
-        des_6 = "Yield"
-        domain += ContinuousVariable(
-            name="yield",
-            description=des_6,
-            bounds=[0.0, 1.0],
-            is_objective=True,
-            maximize=True,
-        )
-
-        return domain
-
-
-class BaumgartnerCrossCouplingEmulator_Yield_Cost(BaumgartnerCrossCouplingEmulator):
-    """Baumgartner Cross Coupling Emulator
-
-    Virtual experiments representing the Aniline Cross-Coupling reaction
-    similar to Baumgartner et al. (2019). Experimental outcomes are based on an
-    emulator that is trained on the experimental data published by Baumgartner et al.
-
-    This is a multiobjective version for optimizing yield and cost simultaneously.
-
-    Parameters
-    ----------
-
-    Examples
-    --------
-    >>> bemul = BaumgartnerCrossCouplingDescriptorEmulator()
-
-    Notes
-    -----
-    This benchmark is based on data from [Baumgartner]_ et al.
-
-    References
-    ----------
-
-    .. [Baumgartner] L. M. Baumgartner et al., Org. Process Res. Dev., 2019, 23, 1594–1601
-       DOI: `10.1021/acs.oprd.9b00236 <https://doi.org/10.1021/acs.oprd.9b00236>`_
-
-    """
-
-    def __init__(self, **kwargs):
+    def __init__(self, domain: Domain, dataset: DataSet, **kwargs):
         super().__init__()
-        self.init_domain = self._domain
-        self.mod_domain = self._domain + ContinuousVariable(
-            name="cost",
-            description="cost in USD of 40 uL reaction",
-            bounds=[0.0, 1.0],
-            is_objective=True,
-            maximize=False,
+        self.domain = domain
+        self.ds = dataset
+        self.normalize = kwargs.get("normalize", True)
+        self.test_size = kwargs.get("test_size", 0.25)
+        self.shuffle = kwargs.get("shuffle", True)
+        self.batch_size = kwargs.get("train_batch_size", 4)
+        self.random_state = kwargs.get("random_state")
+        self.transform = kwargs.get("transform", Transform(self.domain))
+        self.categorical_method = kwargs.get("categorical_method", "one-hot")
+
+        # Run initial setup
+        self.initial_setup()
+
+    def initial_setup(self):
+        # Get data
+        X, y = self.transform.transform_inputs_outputs(
+            self.ds, categorical_method=self.categorical_method
         )
-        self._domain = self.mod_domain
 
-    def _run(self, conditions, **kwargs):
-        # Change to original domain for running predictive model
-        self._domain = self.init_domain
-        conditions, _ = super()._run(conditions=conditions, **kwargs)
+        # Scaling
+        self.input_scaler, self.output_scaler = self._create_scalers(X, y)
+        X = self.input_scaler.transform(X)
+        y = self.output_scaler.transform(y)
 
-        # Calculate costs
-        costs = self._calculate_costs(conditions)
-        conditions[("cost", "DATA")] = costs
+        # Train-test split
+        X_train, X_test, y_train, y_test = train_test_split(
+            X, y, test_size=self.test_size, random_state=self.random_state
+        )
 
-        # Change back to modified domain
-        self._domain = self.mod_domain
-        return conditions, {}
+        # Convert to tensors
+        self.X_train = torch.tensor(X_train).float()
+        self.y_train = torch.tensor(y_train).float()
+        self.X_test = torch.tensor(X_test).float()
+        self.y_test = torch.tensor(y_test).float()
+
+    def train_dataloader(self):
+        ds_train = torch.utils.data.TensorDataset(self.X_train, self.y_train)
+        return torch.utils.data.DataLoader(
+            ds_train, batch_size=self.batch_size, shuffle=self.shuffle
+        )
+
+    def val_dataloader(self):
+        ds_test = torch.utils.data.TensorDataset(self.X_test, self.y_test)
+        return torch.utils.data.DataLoader(ds_test)
 
     @classmethod
-    def _calculate_costs(cls, conditions):
-        catalyst = conditions["catalyst"].values
-        base = conditions["base"].values
-        base_equiv = conditions["base_equivalents"].values
-
-        # Calculate amounts
-        droplet_vol = 40 * 1e-3  # mL
-        mmol_triflate = 0.91 * droplet_vol
-        mmol_anniline = 1.6 * mmol_triflate
-        catalyst_equiv = {
-            "tBuXPhos": 0.0095,
-            "tBuBrettPhos": 0.0094,
-            "AlPhos": 0.0094,
-        }
-        mmol_catalyst = [catalyst_equiv[c] * mmol_triflate for c in catalyst]
-        mmol_base = base_equiv * mmol_triflate
-
-        # Calculate costs
-        cost_triflate = mmol_triflate * 5.91  # triflate is $5.91/mmol
-        cost_anniline = mmol_anniline * 0.01  # anniline is $0.01/mmol
-        cost_catalyst = np.array(
-            [cls._get_catalyst_cost(c, m) for c, m in zip(catalyst, mmol_catalyst)]
-        )
-        cost_base = np.array(
-            [cls._get_base_cost(b, m) for b, m in zip(base, mmol_base)]
-        )
-        tot_cost = cost_triflate + cost_anniline + cost_catalyst + cost_base
-        if len(tot_cost) == 1:
-            tot_cost = tot_cost[0]
-        return tot_cost
+    def from_csv(cls, csv_file, domain, ds, model_dir, **kwargs):
+        """Create a Summit Data Module from a csv file"""
+        ds = DataSet.read_csv(csv_file)
+        return cls(domain, ds, model_dir, **kwargs)
 
     @staticmethod
-    def _get_catalyst_cost(catalyst, catalyst_mmol):
-        catalyst_prices = {
-            "tBuXPhos": 94.08,
-            "tBuBrettPhos": 182.85,
-            "AlPhos": 594.18,
-        }
-        return float(catalyst_prices[catalyst] * catalyst_mmol)
+    def _create_scalers(X, y):
+        input_scaler = StandardScaler().fit(X)
+        output_scaler = StandardScaler().fit(y)
+        return input_scaler, output_scaler
 
-    @staticmethod
-    def _get_base_cost(base, mmol_base):
-        # prices in $/mmol
-        base_prices = {
-            "DBU": 0.03,
-            "BTMG": 1.2,
-            "TMG": 0.001,
-            "TEA": 0.01,
-        }
-        return float(base_prices[base] * mmol_base)
+
+@variational_estimator
+class BayesianRegressor(pl.LightningModule):
+    """A Bayesian Neural Network pytorch lightining module"""
+
+    val_str = "CI acc: {:.2f}, CI upper acc: {:.2f}, CI lower acc: {:.2f}"
+
+    def __init__(self, input_dim, output_dim, train_len, hidden_units=512):
+        super().__init__()
+
+        self.blinear1 = BayesianLinear(input_dim, hidden_units)
+        self.blinear4 = BayesianLinear(hidden_units, output_dim)
+        self.save_hyperparameters("input_dim", "output_dim", "train_len")
+        self.criterion = torch.nn.MSELoss()
+        self.train_len = train_len
+
+    def forward(self, x):
+        x_ = torch.nn.functional.relu(self.blinear1(x))
+        x_ = self.blinear4(x_)
+        return x_
+
+    def training_step(self, batch, batch_idx):
+        X, y = batch
+
+        loss = self.sample_elbo(
+            inputs=X,
+            labels=y,
+            criterion=self.criterion,
+            sample_nbr=3,
+            complexity_cost_weight=1 / self.train_len,
+        )
+        return loss
+
+    # def validation_step(self, batch, batch_idx):
+    #     ic_acc, over_ci_lower, under_ci_upper = self.evaluate_regression(batch)
+
+    #     self.log("val_loss", ic_acc)
+    #     self.log("under_ci_upper", under_ci_upper)
+    #     self.log("over_ci_lower", over_ci_lower)
+
+    #     return ic_acc
+
+    def evaluate_regression(self, batch, samples=100, std_multiplier=1.96):
+        """Evaluate Bayesian Neural Network
+
+        This answers the question "How many correction predictions
+        are in the confidence interval (CI)?" It also spits out the CI.
+
+        Parameters
+        ----------
+        batch : tuple
+            The batch being evaluatd
+        samples : int, optional
+            The number of samples of the BNN for calculating the CI
+        std_multiplier : float, optional
+            The Z-score corresponding with the desired CI. Default is
+            1.96, which corresponds with a 95% CI.
+
+        Returns
+        -------
+        tuple of ic_acc, over_ci_lower, under_ci_upper
+
+        icc_acc is the percentage within the CI.
+
+        """
+
+        X, y = batch
+
+        # Sample
+        preds = [self(X) for i in range(samples)]
+        preds = torch.stack(preds)
+        means = preds.mean(axis=0)
+        stds = preds.std(axis=0)
+
+        # Calculate CI
+        ci_upper, ci_lower = self._calc_ci(means, stds, std_multiplier)
+        ic_acc = (ci_lower <= y) * (ci_upper >= y)
+        ic_acc = ic_acc.float().mean()
+
+        under_ci_upper = (ci_upper >= y).float().mean()
+        over_ci_lower = (ci_lower <= y).float().mean()
+
+        ic_acc = (ci_lower <= y) * (ci_upper >= y)
+        ic_acc = ic_acc.float().mean()
+
+        return ic_acc, over_ci_lower, under_ci_upper
+
+    def _calc_ci(self, means, stds, std_multiplier=1.96):
+        ci_upper = means + (std_multiplier * stds)
+        ci_lower = means - (std_multiplier * stds)
+        return ci_lower, ci_upper
+
+    def configure_optimizers(self):
+        optimizer = torch.optim.SGD(self.parameters(), lr=0.001)
+        return optimizer
+
+    def test_step(self, batch, batch_idx):
+        pass
+
+
+class ANNRegressor(pl.LightningModule):
+    def __init__(self, input_dim, output_dim, hidden_units=512):
+        super().__init__()
+
+        self.linear1 = torch.nn.Linear(input_dim, hidden_units)
+        self.linear2 = torch.nn.Linear(hidden_units, output_dim)
+        self.save_hyperparameters("input_dim", "output_dim")
+        self.criterion = torch.nn.MSELoss()
+
+    def forward(self, x):
+        x_ = self.linear1(x)
+        x_ = torch.nn.functional.relu(x_)
+        return self.linear2(x_)
+
+    def training_step(self, batch, batch_idx):
+        X, y = batch
+        y_hat = self(X)
+        loss = self.criterion(y_hat, y)
+        self.log("train_loss", loss)
+        return loss
+
+    def validation_step(self, batch, batch_idx):
+        X, y = batch
+        y_hat = self(X)
+        loss = self.criterion(y_hat, y)
+        self.log("val_loss", loss)
+
+    def configure_optimizers(self):
+        optimizer = torch.optim.SGD(self.parameters(), lr=0.001)
+        return optimizer
+
+
+def create_domain():
+    domain = Domain()
+    domain += ContinuousVariable(
+        name="temperature",
+        description="reaction temperature in celsius",
+        bounds=[30, 100],
+    )
+    domain += ContinuousVariable(
+        name="flowrate_a", description="flowrate of reactant a", bounds=[1, 100]
+    )
+
+    domain += ContinuousVariable(
+        name="flowrate_b", description="flowrate of reactant b", bounds=[1, 100]
+    )
+
+    domain += ContinuousVariable(
+        name="yield",
+        description="yield of reaction",
+        bounds=[0, 100],
+        is_objective=True,
+        maximize=True,
+    )
+    return domain
+
+
+def create_dataset(domain, n_samples=100, random_seed=100):
+    rng = default_rng(random_seed)
+    n_features = len(domain.input_variables)
+    inputs = rng.standard_normal(size=(n_samples, n_features))
+    inputs *= [-5, 6, 0.1]
+    output = np.sum(inputs ** 2, axis=1)
+    data = np.append(inputs, np.atleast_2d(output).T, axis=1)
+    columns = [v.name for v in domain.input_variables] + [
+        domain.output_variables[0].name
+    ]
+    return DataSet(data, columns=columns)
+
+
+def main():
+    # Get data
+    domain = create_domain()
+    dataset = create_dataset(domain, n_samples=500)
+
+    # Setup and train model
+    n_features = len(domain.input_variables)
+    regressor = ANNRegressor(n_features, 1)
+    exp = ExperimentalEmulator("test_ann_model", domain, dataset, regressor=regressor)
+    exp.train(max_epochs=10)
+    exp.parity_plot()
+    plt.show()
+
+
+if __name__ == "__main__":
+    main()
